fix(pipelines): bigint ids for file-drop sources + robust JDBC connect in E2E
Date-derived shipment/return ids (yyyymmddNNN) overflow INT32 and DBR's
ANSI mode fails the cast (local Spark tolerated it). The SCD2 proof now
uses the 3-arg DriverManager.getConnection overload so generated passwords
with special characters never touch the connection string.

P2 exit criteria verified on Databricks 2026-07-04 (scripts/run_p2_e2e.py):
all 9 checks PASS — gold fed from both sources (fact_sales=15103,
agg_distributor_shipments=15) and SCD2 closed-old/new-current confirmed.

+++ src/lakeforge/transform/silver.py
@@ -36,7 +36,8 @@ UPSERT_KEYS = {
 # Casts applied to the file-based sources (Auto Loader CSV reads as strings).
 FILE_CASTS: dict[str, dict[str, str]] = {
     "shipments": {
-        "shipment_id": "int",
+        # File-drop ids are date-derived (yyyymmddNNN) and overflow INT32.
+        "shipment_id": "bigint",
         "order_id": "int",
         "distributor": "string",
         "ship_date": "date",
@@ -44,7 +45,7 @@ FILE_CASTS: dict[str, dict[str, str]] = {
         "warehouse": "string",
     },
     "returns": {
-        "return_id": "int",
+        "return_id": "bigint",
         "order_id": "int",
         "product_sku": "string",
         "return_date": "date",
